+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 v0.4.0 (in development)
 -----------------------
+- Added tokio-based asynchronous functionality:
+    - `AsyncJsonLinesReader`
+    - `AsyncJsonLinesWriter`
 
 v0.3.0 (2022-10-30)
 -------------------